Editor: Move editorstack widget to its own module
