+++ spyder/plugins/editor/widgets/editorstack/__init__.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8 -*-
+# -----------------------------------------------------------------------------
+# Copyright (c) 2009- Spyder Project Contributors
+#
+# Distributed under the terms of the MIT License
+# (see spyder/__init__.py for details)
+# -----------------------------------------------------------------------------
+
+
+"""
+spyder.plugins.editor.widgets.editorstack
+=========================================
+
+Main widget handler of CodeEditors.
+"""
+
+from .editorstack import EditorStack
