+++ spyder/plugins/editor/widgets/editorstack/editorstack_helpers.py
